pyproject.toml
[project]
name = "discord-reels-reposter"
version = "0.1.0"
description = "Discord bot that re-embeds Instagram Reels and TikTok videos as native videos"
requires-python = ">=3.9"
dependencies = [
    "discord.py>=2.3.0",
    "yt-dlp>=2023.10.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.0.0",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
discord-reels-reposter = "bot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cogs", "utils"]
